Added the command "/autoloot buy "item name" #", it will buy up to # of the item specified on the merchant you have opened.
Fixed a bug that would cause the plugin to crash if your leader is in another zone, you don't have a master looter, and the advanced loot window pops open.

Made it so that the plugin will check if you are hovering and not to anything if that is the case.  Sumon claimed it was causing a crash, I couldn't reproduce it but it didn't hurt to make the change.